backend/requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.5
django-cors-headers==4.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
idna==3.4
install==1.3.5
mysqlclient==2.2.0
Pillow==10.1.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
